// scripts/src/common/codegen/Counter.client.ts
/**
* This file was automatically generated by @cosmwasm/ts-codegen@0.35.3.
* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
*/

import { CosmWasmClient, SigningCosmWasmClient, ExecuteResult } from "@cosmjs/cosmwasm-stargate";
import { Coin, StdFee } from "@cosmjs/amino";
import { InstantiateMsg, ExecuteMsg, ActionType, Uint128, QueryMsg, MigrateMsg, Addr, Config, ArrayOfTupleOfAddrAndUint128 } from "./Counter.types";
export interface CounterReadOnlyInterface {
  contractAddress: string;
  queryCounters: () => Promise<ArrayOfTupleOfAddrAndUint128>;
  queryConfig: () => Promise<Config>;
}
export class CounterQueryClient implements CounterReadOnlyInterface {
  client: CosmWasmClient;
  contractAddress: string;

  constructor(client: CosmWasmClient, contractAddress: string) {
    this.client = client;
    this.contractAddress = contractAddress;
    this.queryCounters = this.queryCounters.bind(this);
    this.queryConfig = this.queryConfig.bind(this);
  }

  queryCounters = async (): Promise<ArrayOfTupleOfAddrAndUint128> => {
    return this.client.queryContractSmart(this.contractAddress, {
      query_counters: {}
    });
  };
  queryConfig = async (): Promise<Config> => {
    return this.client.queryContractSmart(this.contractAddress, {
      query_config: {}
    });
  };
}
export interface CounterInterface extends CounterReadOnlyInterface {
  contractAddress: string;
  sender: string;
  createCounter: (fee?: number | StdFee | "auto", memo?: string, _funds?: Coin[]) => Promise<ExecuteResult>;
  updateCounter: ({
    actionType,
    value
  }: {
    actionType: ActionType;
    value: Uint128;
  }, fee?: number | StdFee | "auto", memo?: string, _funds?: Coin[]) => Promise<ExecuteResult>;
  resetCounter: (fee?: number | StdFee | "auto", memo?: string, _funds?: Coin[]) => Promise<ExecuteResult>;
}
export class CounterClient extends CounterQueryClient implements CounterInterface {
  client: SigningCosmWasmClient;
  sender: string;
  contractAddress: string;

  constructor(client: SigningCosmWasmClient, sender: string, contractAddress: string) {
    super(client, contractAddress);
    this.client = client;
    this.sender = sender;
    this.contractAddress = contractAddress;
    this.createCounter = this.createCounter.bind(this);
    this.updateCounter = this.updateCounter.bind(this);
    this.resetCounter = this.resetCounter.bind(this);
  }

  createCounter = async (fee: number | StdFee | "auto" = "auto", memo?: string, _funds?: Coin[]): Promise<ExecuteResult> => {
    return await this.client.execute(this.sender, this.contractAddress, {
      create_counter: {}
    }, fee, memo, _funds);
  };
  updateCounter = async ({
    actionType,
    value
  }: {
    actionType: ActionType;
    value: Uint128;
  }, fee: number | StdFee | "auto" = "auto", memo?: string, _funds?: Coin[]): Promise<ExecuteResult> => {
    return await this.client.execute(this.sender, this.contractAddress, {
      update_counter: {
        action_type: actionType,
        value
      }
    }, fee, memo, _funds);
  };
  resetCounter = async (fee: number | StdFee | "auto" = "auto", memo?: string, _funds?: Coin[]): Promise<ExecuteResult> => {
    return await this.client.execute(this.sender, this.contractAddress, {
      reset_counter: {}
    }, fee, memo, _funds);
  };
}